general: get rid of catchall route
