--- pages/[[...all]].ts
@@ -1,1 +1,0 @@
-export { default } from '../src/components/App/App';
