+++ pages/_app.tsx
@@ -13,8 +13,47 @@ import { doShortenerRedirect } from '../src/services/helpers';
 import { PROJECT_ID, PROJECT_NAME } from '../src/services/project';
 import { GoogleAnalytics } from './api/google-analytics';
 import { Umami } from './api/umami';
+import { SnackbarProvider } from '../src/components/utils/SnackbarContext';
+import { UserSettingsProvider } from '../src/components/utils/UserSettingsContext';
+import {
+  MapStateProvider,
+  View,
+} from '../src/components/utils/MapStateContext';
+import { OsmAuthProvider } from '../src/components/utils/OsmAuthContext';
+import { EditDialogProvider } from '../src/components/FeaturePanel/helpers/EditDialogContext';
+import Map from '../src/components/Map/Map';
+import { TitleAndMetaTags } from '../src/helpers/TitleAndMetaTags';
+import {
+  getInitialFeature,
+  getInitialMapView,
+  getMapViewFromHash,
+} from '../src/components/App/helpers';
+import { FeatureProvider } from '../src/components/utils/FeatureContext';
+import { StarsProvider } from '../src/components/utils/StarsContext';
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { Loading } from '../src/components/App/Loading';
+import { SearchBox } from '../src/components/SearchBox/SearchBox';
+import { HomepagePanel } from '../src/components/HomepagePanel/HomepagePanel';
+import { setIntlForSSR, t } from '../src/services/intl';
+import { Feature } from '../src/services/types';
+import { ResponsiveFeaturePanel } from '../src/components/FeaturePanel/ResponsiveFeaturePanel';
+import { Climbing } from '../src/components/Climbing/Climbing';
 
-type Props = AppProps & EmotionCacheProviderProps;
+const URL_NOT_FOUND_TOAST = {
+  message: t('url_not_found_toast'),
+  severity: 'warning' as const,
+};
+
+const reactQueryClient = new QueryClient();
+
+type OwnProps = {
+  userThemeCookie: string;
+  featureFromRouter: Feature | '404' | null;
+  initialMapView: View;
+  cookies: Record<string, string>;
+};
+
+type Props = AppProps & EmotionCacheProviderProps & OwnProps;
 
 export default class MyApp extends App<Props> {
   componentDidMount() {
@@ -27,9 +66,16 @@ export default class MyApp extends App<Props> {
   }
 
   render() {
-    const { Component, pageProps, emotionCache } = this.props as any;
-    const { userThemeCookie } = pageProps;
-    const isOpenClimbing = PROJECT_ID === 'openclimbing';
+    const {
+      Component,
+      pageProps,
+      emotionCache,
+      userThemeCookie,
+      initialMapView,
+      cookies,
+      featureFromRouter,
+    } = this.props;
+    const mapView = getMapViewFromHash() || initialMapView;
 
     return (
       <>
@@ -45,11 +91,42 @@ export default class MyApp extends App<Props> {
             {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
             <CssBaseline />
             {/* eslint-disable-next-line react/jsx-props-no-spreading */}
-            <Component {...pageProps} />
+
+            <SnackbarProvider initialToast={undefined}>
+              <UserSettingsProvider>
+                <FeatureProvider
+                  featureFromRouter={
+                    featureFromRouter === '404' ? null : featureFromRouter
+                  }
+                  cookies={cookies}
+                >
+                  <MapStateProvider initialMapView={mapView}>
+                    <OsmAuthProvider cookies={cookies}>
+                      <StarsProvider>
+                        <EditDialogProvider /* TODO supply router.query */>
+                          <QueryClientProvider client={reactQueryClient}>
+                            <Loading />
+                            <SearchBox />
+                            <ResponsiveFeaturePanel />
+                            <HomepagePanel />
+                            <Climbing />
+                            <Map />
+                            <TitleAndMetaTags />
+
+                            <Component {...pageProps} />
+                          </QueryClientProvider>
+                        </EditDialogProvider>
+                      </StarsProvider>
+                    </OsmAuthProvider>
+                  </MapStateProvider>
+                </FeatureProvider>
+              </UserSettingsProvider>
+            </SnackbarProvider>
+
             <GlobalStyle />
           </UserThemeProvider>
         </AppCacheProvider>
-        {isOpenClimbing && (
+        {PROJECT_ID === 'openclimbing' && (
           <>
             <GoogleAnalytics />
             <Umami />
@@ -65,13 +142,30 @@ MyApp.getInitialProps = async ({ ctx, Component }) => {
     return { pageProps: undefined };
   }
 
+  await setIntlForSSR(ctx); // needed for lang urls like /es/node/123
+
   const pageProps = await Component.getInitialProps?.(ctx);
   const { userTheme } = nextCookies(ctx);
 
+  const cookies = nextCookies(ctx);
+
+  console.log(ctx.pathname);
+  const featureFromRouter = await getInitialFeature(ctx);
+  if (ctx.res) {
+    if (featureFromRouter === '404' || featureFromRouter?.error === '404') {
+      ctx.res.statusCode = 404;
+    } else if (featureFromRouter?.error) {
+      ctx.res.statusCode = 500;
+    }
+  }
+
+  const initialMapView = await getInitialMapView(ctx);
+
   return {
-    pageProps: {
-      ...pageProps,
-      userThemeCookie: userTheme,
-    },
+    featureFromRouter,
+    initialMapView,
+    cookies,
+    userThemeCookie: userTheme,
+    pageProps,
   };
 };
