+++ pages/index.ts
@@ -1,0 +1,2 @@
+// HomepagePanel is rendered conditionally, see useFeatureContext()
+export default () => null;
