+++ pages/way/[id].ts
@@ -1,0 +1,2 @@
+// The FeaturePanel is rendered in _app (skeleton shows it before URL change, see ARCHITECTURE.md for more info)
+export default () => null;
